move SNL5MW test dependencies out of OWENS and into file.
